GUI: moved the classes AbstractIodeObjectWidget and QIodeObjectWidget to the new header file tabs/tab_database_abstract.h
