+++ gui/tabs/tab_database.h
@@ -1,18 +1,6 @@
 #pragma once
 
-#include <QLabel>
-#include <QObject>
-#include <QWidget>
-#include <QSpinBox>
-#include <QShortcut>
-#include <QLineEdit>
-#include <QGridLayout>
-#include <QPushButton>
-#include <QSizePolicy>
-#include <QStringList>
-
-#include "tab_abstract.h"
-#include "custom_widgets/tab_widget_abstract.h"
+#include "tab_database_abstract.h"
 
 #include "iode_objs/models/comments_model.h"
 #include "iode_objs/models/equations_model.h"
@@ -29,289 +17,6 @@
 #include "iode_objs/views/scalars_view.h"
 #include "iode_objs/views/tables_view.h"
 #include "iode_objs/views/variables_view.h"
-
-const static QString prefixUnsavedDatabase = "Unsaved";
-
-
-class AbstractIodeObjectWidget: public AbstractTabWidget
-{
-    Q_OBJECT
-
-protected:
-    EnumIodeType iodeType;
-    QDir projectDir;
-
-    QGridLayout* layout;
-    QLineEdit* lineEdit_filter;
-    QPushButton* pushButton_filter;
-    QPushButton* pushButton_print;
-    QPushButton* pushButton_add;
-
-signals:
-    void tabDatabaseModified(const EnumIodeType iodeType, const bool modified);
-
-public:
-    AbstractIodeObjectWidget(const EnumIodeType iodeType, QIodeAbstractTabWidget* parent) : 
-        AbstractTabWidget((EnumIodeFile) iodeType, parent), projectDir(QDir::homePath()), iodeType(iodeType)
-    {
-        this->setObjectName(QString::fromUtf8("widget_iode_obj"));
-
-        // layout
-        layout = new QGridLayout(this);
-        layout->setObjectName(QString::fromUtf8("layout"));
-        layout->setContentsMargins(0, 0, 0, 0);
-
-        // filter 
-        lineEdit_filter = new QLineEdit();
-        lineEdit_filter->setObjectName(QString::fromUtf8("lineEdit_filter"));
-        lineEdit_filter->setPlaceholderText("filter pattern here");
-        QSizePolicy sizePolicyFilter(QSizePolicy::Expanding, QSizePolicy::Fixed);
-        sizePolicyFilter.setHorizontalStretch(0);
-        sizePolicyFilter.setVerticalStretch(0);
-        sizePolicyFilter.setHeightForWidth(lineEdit_filter->sizePolicy().hasHeightForWidth());
-        lineEdit_filter->setSizePolicy(sizePolicyFilter);
-        lineEdit_filter->setMinimumSize(QSize(200, 0));
-        layout->addWidget(lineEdit_filter, 0, 0, Qt::AlignLeft);
-
-        pushButton_filter = new QPushButton("Filter");
-        pushButton_filter->setObjectName(QString::fromUtf8("pushButton_filter"));
-        layout->addWidget(pushButton_filter, 0, 1, Qt::AlignLeft);
-
-        // spacer
-        QSpacerItem* horizontalSpacer = new QSpacerItem(828, 20, QSizePolicy::Expanding, QSizePolicy::Minimum);
-        layout->addItem(horizontalSpacer, 0, 2);
-
-        // print button
-        pushButton_print = new QPushButton("Print");
-        pushButton_print->setObjectName(QString::fromUtf8("pushButton_print"));
-        layout->addWidget(pushButton_print, 0, 3, Qt::AlignLeft);
-
-        // add button
-        pushButton_add = new QPushButton("Add " + QString::fromStdString(vIodeTypes[iodeType]));
-        pushButton_add->setObjectName(QString::fromUtf8("pushButton_add"));
-        QSizePolicy sizePolicyAdd(QSizePolicy::Fixed, QSizePolicy::Fixed);
-        sizePolicyAdd.setHorizontalStretch(0);
-        sizePolicyAdd.setVerticalStretch(0);
-        sizePolicyAdd.setHeightForWidth(pushButton_add->sizePolicy().hasHeightForWidth());
-        pushButton_add->setSizePolicy(sizePolicyAdd);
-        layout->addWidget(pushButton_add, 0, 4, Qt::AlignRight);
-    }
-
-    ~AbstractIodeObjectWidget()
-    {
-        delete lineEdit_filter;
-        delete pushButton_filter;
-        delete pushButton_add;
-        delete layout;
-    }
-
-    /**
-     * @brief Return whether or not the tooltip is associated with a tab representing an unsaved KDB.
-     * 
-     * @param QString Tooltip of the tab. 
-     * @return bool
-     */
-    bool isUnsavedDatabase() const
-    {
-        QString filepath = getFilepath();
-        return filepath.isEmpty() || filepath == QString(I_DEFAULT_FILENAME);
-    }
-
-    QString getTabText() const
-    {
-        if(isUnsavedDatabase())
-        {
-            QString ext = QString::fromStdString(v_binary_ext[fileType]);
-            // Note: the * is to tell that the content of the KDB has not been saved in file
-            return tabPrefix[fileType] + QString(I_DEFAULT_FILENAME) + "." + ext + "*";
-        }
-        else
-            return AbstractTabWidget::getTabText();
-    }
-
-    QString getTooltip() const
-    {
-        if(isUnsavedDatabase())
-            return prefixUnsavedDatabase + " " + QString::fromStdString(vIodeTypes[iodeType]) + " Database";
-        else
-            return AbstractTabWidget::getTooltip();
-    }
-
-    bool updateFilepath(const QString& filepath) override
-    {
-        if(checkNewFilepath(filepath))
-        {
-            set_kdb_filename(K_WS[fileType], filepath.toStdString());
-            modified = false;
-            emit tabDatabaseModified(iodeType, false);
-            return true;
-        }
-        else
-            return false;
-    }
-
-    void reset()
-    {
-        clearKDB();
-        resetFilter();
-    }
-
-    virtual void setProjectDir(const QDir& projectDir)
-    {
-        this->projectDir = projectDir;
-        reset();
-    }
-
-    virtual void clearKDB() = 0;
-    virtual void resetFilter() = 0;
-    virtual QStringList getSelectedObjectsNames() const = 0;
-    virtual void computeHash(const bool before=false) = 0;
-
-public slots:
-    void setModified(bool modified) override
-    {
-        this->modified = modified;
-        emit tabDatabaseModified(iodeType, modified); 
-    }
-
-    void databaseModified()
-    {
-        setModified(true);
-    }
-};
-
-
-template <class M, class V> class QIodeObjectWidget: public AbstractIodeObjectWidget
-{
-protected:
-    M* objmodel;
-    V* tableview;
-
-    QShortcut* shortcutPrint;
-    QShortcut* shortcutAdd;
-
-public:
-    QIodeObjectWidget(EnumIodeType iodeType, QIodeAbstractTabWidget* parent) 
-        : AbstractIodeObjectWidget(iodeType, parent)
-    {
-        // view table
-        tableview = new V(parent);
-        tableview->setObjectName(QString::fromUtf8("tableview"));
-        tableview->setGeometry(QRect(10, 43, 950, 560));
-
-        // model table
-        objmodel = new M(tableview);
-        tableview->setModel(objmodel);
-
-        // IODE objects names filter
-        tableview->setFilterLineEdit(lineEdit_filter);
-        
-        // shortcuts
-        shortcutPrint = new QShortcut(QKeySequence(Qt::CTRL | Qt::Key_P), this);
-        shortcutAdd = new QShortcut(QKeySequence(Qt::Key_Insert), this);
-
-        // NOTE: By default, shortcuts are defined at the main Window level. 
-        //       Thus, a shortcut of a (combination of) key(s) may override the expected behavior 
-        //       from another widget dealing with the same (combination of) key(s). 
-        //       'setContext(Qt::WidgetWithChildrenShortcut)' makes sure that the shortcut does 
-        //       not propagate to other widgets.
-        shortcutPrint->setContext(Qt::WidgetWithChildrenShortcut);
-        shortcutAdd->setContext(Qt::WidgetWithChildrenShortcut);
-
-        // signals - slots
-        connect(lineEdit_filter, &QLineEdit::returnPressed, tableview, &V::filter);
-        connect(pushButton_filter, &QPushButton::clicked, tableview, &V::filter);
-
-        connect(pushButton_print, &QPushButton::clicked, tableview, &V::print);
-        connect(shortcutPrint, &QShortcut::activated, tableview, &V::print);
-
-        connect(pushButton_add, &QPushButton::clicked, tableview, &V::new_obj);
-        connect(shortcutAdd, &QShortcut::activated, tableview, &V::new_obj);
-
-        connect(tableview, &AbstractTableView::showObjsRequest, parent, &QIodeAbstractTabWidget::showObjectsList);
-
-        // insert table to layout
-        // -1 -> span over all rows/columns
-        layout->addWidget(tableview, 1, 0, 1, -1);
-    }
-
-    ~QIodeObjectWidget() 
-    {
-        delete objmodel;
-        delete tableview;
-        delete shortcutPrint;
-        delete shortcutAdd;
-    }
-
-    M* get_model() const 
-    {
-        return objmodel;
-    }
-
-    V* get_view() const
-    {
-        return tableview;
-    }
-
-    QString getFilepath() const
-    {
-        return objmodel->getFilepath();
-    }
-
-    void clearKDB()
-    {
-        objmodel->clearKDB();
-        tableview->update();
-    }
-
-    void resetFilter()
-    {
-        lineEdit_filter->setText("");
-        tableview->filter();
-    }
-
-    void update()
-    {
-        tableview->update();
-    }
-
-    void setup()
-    {
-        tableview->setup();
-    }
-
-    bool load_(const QString& filepath, const bool forceOverwrite)
-    {
-        return objmodel->load(filepath, forceOverwrite);
-    }
-
-    QString save()
-    {
-        if(isUnsavedDatabase()) 
-            return saveAs();
-        else
-        {
-            QString filepath = objmodel->save(projectDir);
-            if(!filepath.isEmpty()) setModified(false);
-            return filepath;
-        }
-    }
-
-    QString saveAs_()
-    {
-        return objmodel->saveAs(projectDir);
-    }
-
-    QStringList getSelectedObjectsNames() const 
-    {
-        return tableview->getSelectedObjectsNames();
-    }
-
-    void computeHash(const bool before=false)
-    {
-        objmodel->computeHash(before);
-    }
-};
 
 
 /**
